Refactor utils to a separate package
