+++ packages/utils/README.md
@@ -1,0 +1,11 @@
+# `utils`
+
+> TODO: description
+
+## Usage
+
+```
+const utils = require('@oasisdex/utils');
+
+// TODO: DEMONSTRATE API
+```
