+++ packages/utils/src/index.ts
@@ -1,0 +1,1 @@
+export { amountFromWei } from './utils';
