+++ packages/utils/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "../../tsconfig-base",
+  "compilerOptions": {
+    "outDir": "lib",
+    "rootDir": "."
+  },
+  "include": ["./src/**/*"]
+}
