+++ packages/web3-context/package.json
@@ -21,6 +21,7 @@
     "tsc-watch": "^4.2.9"
   },
   "dependencies": {
+    "@oasisdex/utils": "^0.0.1",
     "@oasisdex/connectors": "^0.0.14",
     "@web3-react/abstract-connector": "^6.0.7",
     "@web3-react/core": "^6.1.1",
