+++ tsconfig.json
@@ -5,6 +5,6 @@
     { "path": "./packages/connectors" },
     { "path": "./packages/transactions" },
     { "path": "./packages/web3-context" },
+    { "path": "./packages/utils" }
   ]
 }
-
